+++ packages/e2e-tests/config/routes.ts
@@ -1,9 +1,9 @@
 export const routes = {
   login: '/',
-  dashboard: '#/dashboard',
+  dashboard: '/#/dashboard',
   patients: {
-    all: '#/patients/all',
-    patientDetails: '#/patients/all/*',
+    all: '/#/patients/all',
+    patientDetails: '/#/patients/all/*',
   },
 };
