+++ packages/e2e-tests/tests/setup/auth.setup.ts
@@ -1,13 +1,17 @@
 import { test as setup } from '../../fixtures/baseFixture';
 import path from 'path';
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
 
 setup('authenticate', async ({ loginPage }) => {
   await loginPage.goto();
-  await loginPage.login(process.env.TEST_EMAIL, process.env.TEST_PASSWORD);
+
+  const email = process.env.TEST_EMAIL;
+  const password = process.env.TEST_PASSWORD;
+
+  if (!email || !password) {
+    throw new Error('TEST_EMAIL and TEST_PASSWORD environment variables must be set');
+  }
+
+  await loginPage.login(email, password);
 
   // Ensure state is persisted
   await loginPage.page.waitForTimeout(1000);
