+++ packages/fake-data/package.json
@@ -3,7 +3,6 @@
   "version": "2.35.0",
   "private": true,
   "description": "TODO",
-  "type": "module",
   "main": "dist/cjs/index.js",
   "module": "dist/mjs/index.js",
   "exports": {
@@ -19,9 +18,9 @@
       "import": "./dist/mjs/fake/index.js",
       "require": "./dist/cjs/fake/index.js"
     },
-    "./fake/fakeRequest/fakeCreatePatient": {
-      "import": "./dist/mjs/fake/fakeRequest/fakeCreatePatient.js",
-      "require": "./dist/cjs/fake/fakeRequest/fakeCreatePatient.js"
+    "./fake/fakeRequest/createPatient": {
+      "import": "./dist/mjs/fake/fakeRequest/createPatient.js",
+      "require": "./dist/cjs/fake/fakeRequest/createPatient.js"
     }
   },
   "homepage": "https://github.com/beyondessential/tamanu.git#readme",
