+++ packages/fake-data/src/fake/fakeRequest/index.ts
@@ -1,1 +1,1 @@
-export * from './fakeCreatePatient.js';
+export * from './createPatient.js';
