+++ packages/fake-data/tsconfig.json
@@ -8,10 +8,11 @@
     "experimentalDecorators": true,
     "isolatedModules": false,
     "noEmit": true,
+    "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true,
-    "moduleResolution": "NodeNext",
-    "module": "NodeNext",
     "skipLibCheck": true,
-    "target": "es2020"
+    "target": "es2020",
+    "moduleResolution": "bundler",
+    "module": "ESNext"
   }
 }
