+++ packages/shared/package.json
@@ -66,8 +66,8 @@
       "import": "./dist/esm/audit/*.js"
     },
     "./*": {
-      "require": "./dist/cjs/*/index.js",
-      "import": "./dist/esm/*/index.js"
+      "import": "./dist/esm/*/index.js",
+      "require": "./dist/cjs/*/index.js"
     }
   },
   "homepage": "https://github.com/beyondessential/tamanu.git#readme",
